% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Accessors.R
\docType{methods}
\name{Accessors}
\alias{$,OriginPeriod-method}
\alias{$<-,OriginPeriod-method}
\alias{Accessors}
\alias{[,OriginPeriod,ANY-method}
\alias{[,OriginPeriod,character-method}
\alias{[<-,OriginPeriod,ANY,ANY,OriginPeriod-method}
\title{Accessors}
\usage{
\S4method{[}{OriginPeriod,ANY}(x, i)

\S4method{[}{OriginPeriod,character}(x, i)

\S4method{[}{OriginPeriod,ANY,ANY,OriginPeriod}(x, i, j, ...) <- value

\S4method{$}{OriginPeriod}(x, name)

\S4method{$}{OriginPeriod}(x, name) <- value
}
\arguments{
\item{x}{OriginPeriod object}

\item{i}{Which OriginPeriod element}

\item{j}{Not used}

\item{...}{Not used}

\item{value}{OriginPeriod object}

\item{name}{Slot to alter}
}
\description{
Accessors
}

