% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{Concatenation}
\alias{Concatenation}
\alias{c,OriginPeriod-method}
\alias{rbind.OriginPeriod}
\title{Concatenation}
\usage{
rbind.OriginPeriod(..., deparse.level = 1)

\S4method{c}{OriginPeriod}(x, ..., recursive = FALSE)
}
\arguments{
\item{deparse.level}{I don't know what this is}

\item{...}{OriginPeriod objects to bind}

\item{x}{OriginPeriod object}

\item{recursive}{logical.  If \code{recursive = TRUE}, the function
    recursively descends through lists (and pairlists) combining all
    their elements into a vector.}
}
\value{
OriginPeriod object

OriginPeriod object
}
\description{
Concatenation
}

