% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Concatenation.R
\docType{methods}
\name{Grow}
\alias{Grow}
\alias{Grow,OriginPeriod-method}
\title{Grow}
\usage{
Grow(object, ...)

\S4method{Grow}{OriginPeriod}(object, Length)
}
\arguments{
\item{object}{OriginPeriod object}

\item{...}{Additional arguments}

\item{Length}{Number of elements to add}
}
\description{
Grow will expand the length of an OriginPeriod object by an arbitrary length.
}

