% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Constructors.R
\docType{methods}
\name{OriginPeriod}
\alias{OriginPeriod}
\alias{OriginPeriod,ANY-method}
\title{OriginPeriod}
\usage{
OriginPeriod(StartDate, EndDate, Period, ...)

\S4method{OriginPeriod}{ANY}(StartDate, EndDate, Period, Moniker, Type,
  Verbose = FALSE, NumPeriods, StartMonth, StartDay)
}
\arguments{
\item{StartDate}{vector of start dates}

\item{EndDate}{vector of end dates}

\item{Period}{The period between the start and end dates}

\item{...}{Additional parameters}

\item{Moniker}{Character vector of names for each element}

\item{Type}{"Accident year", "Report year", etc.}

\item{Verbose}{Display warnings?}

\item{NumPeriods}{How many periods?}

\item{StartMonth}{Starting month of the period}

\item{StartDay}{Starting day of the period}
}
\description{
Construct an OriginPeriod object
}
\details{
An OriginPeriod object may be constructed in a number of ways.

1. Provide start and end dates
}

