% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Persistence.R
\docType{methods}
\name{write.excel}
\alias{write.excel}
\alias{write.excel,OriginPeriod,character,logical-method}
\title{write.excel}
\usage{
write.excel(object, file, overwrite, ...)

\S4method{write.excel}{OriginPeriod,character,logical}(object, file,
  overwrite = FALSE, sheetName)
}
\arguments{
\item{object}{Object to write to Excel}

\item{file}{Character string with the name of the file}

\item{overwrite}{Overwrite the file if it exists?}

\item{...}{Additional arguments}

\item{sheetName}{Name of sheet where the object will be written}
}
\description{
write.excel
}

