% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{Accessors}
\alias{$,OriginPeriod-method}
\alias{$<-,OriginPeriod-method}
\alias{Accessors}
\alias{[,OriginPeriod,character-method}
\alias{[.OriginPeriod}
\alias{[<-,OriginPeriod,ANY,ANY,OriginPeriod-method}
\title{Accessors}
\usage{
\method{[}{OriginPeriod}(x, i)

\S4method{[}{OriginPeriod,character}(x, i)

\S4method{[}{OriginPeriod,ANY,ANY,OriginPeriod}(x, i, j, ...) <- value

\S4method{$}{OriginPeriod}(x, name)

\S4method{$}{OriginPeriod}(x, name) <- value
}
\arguments{
\item{x}{OriginPeriod object}

\item{value}{OriginPeriod object}

\item{i,j}{Indices}

\item{...}{Additional arguments (ignored)}

\item{name}{
    A literal character string or a \link{name} (possibly \link{backtick}
    quoted).  For extraction, this is normally (see under
    \sQuote{Environments}) partially matched to the \code{\link{names}}
    of the object.
  }
}
\description{
Accessors
}

