% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{Comparison}
\alias{!=,OriginPeriod,OriginPeriod-method}
\alias{<,OriginPeriod,OriginPeriod-method}
\alias{<=,OriginPeriod,OriginPeriod-method}
\alias{==,OriginPeriod,OriginPeriod-method}
\alias{>,OriginPeriod,OriginPeriod-method}
\alias{>=,OriginPeriod,OriginPeriod-method}
\alias{Comparison}
\title{Comparison}
\usage{
\S4method{==}{OriginPeriod,OriginPeriod}(e1, e2)

\S4method{!=}{OriginPeriod,OriginPeriod}(e1, e2)

\S4method{>}{OriginPeriod,OriginPeriod}(e1, e2)

\S4method{>=}{OriginPeriod,OriginPeriod}(e1, e2)

\S4method{<}{OriginPeriod,OriginPeriod}(e1, e2)

\S4method{<=}{OriginPeriod,OriginPeriod}(e1, e2)
}
\arguments{
\item{e1}{OriginPeriod vector}

\item{e2}{OriginPeriod vector}
}
\description{
Comparison
}

