% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Concatenation.R
\docType{methods}
\name{Concatenation}
\alias{Concatenation}
\alias{c,OriginPeriod-method}
\alias{rbind.OriginPeriod}
\title{Concatenation}
\usage{
rbind.OriginPeriod(..., deparse.level = 1)

\S4method{c}{OriginPeriod}(x, ..., recursive = FALSE)
}
\arguments{
\item{...}{OriginPeriod objects to bind}

\item{deparse.level}{I don't know what this is}

\item{x}{OriginPeriod object}

\item{recursive}{Not used}
}
\value{
OriginPeriod object

OriginPeriod object
}
\description{
Concatenation
}

