% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Conversion.R
\name{Conversion}
\alias{Conversion}
\alias{as.data.frame.OriginPeriod}
\title{Conversion}
\usage{
\method{as.data.frame}{OriginPeriod}(x, ...)
}
\arguments{
\item{x}{OriginPeriod object}

\item{...}{Additional arguments}
}
\description{
Converts an OriginPeriod object to a data frame
}

